Update submodule configuration for RBE549-S25-S01.wiki
Following a few SO threads, refer to discussion post I also posted #41, I have updated the submodule configuration for the RBE549-S25-S01.wiki submodule. The branch has been set to master and the update has been set to rebase. This will allow the submodule to be updated with the latest changes from the master branch of the submodule repository.

Signed-off-by: Neehal Sharrma <nasharrma@wpi.edu>