Add Blender render pipeline
Modified the pipeline to run through the JSON file and spawn objects based on the extracted positions and rotations. The pipeline now supports spawning vehicles and traffic-related objects based on the object type specified in the JSON file. The pipeline also renders the scene and saves the rendered image to the specified directory.

Signed-off-by: Neehal Sharrma <nasharrma@wpi.edu>