(Incomplete) Add Wrapper and model pipeline content
I'm facing issues with the home stretch of just tying everything together but I think I have the rest down. Be advised that the PyTorch hub was giving me major issues with the ZoeDepth model so I tried to download it manually and place it in the Networks folder. I also had to change the path for the YOLOv11 model to point to a local file instead of the repo.

Signed-off-by: Neehal Sharrma <nasharrma@wpi.edu>